changed temp/humidity to floats on table, added run.rb file

+++ db/migrate/004_create_weathers.rb
@@ -2,8 +2,8 @@ class CreateWeathers < ActiveRecord::Migration
     def change 
         create table :weathers do |t|
             t.string :location
-            t.integer :temperature 
-            t.integer :humidity
+            t.float :temperature 
+            t.float:humidity
             t.integer :location_id
         end
     end
